moved vtk parser inside utils folder
